// src/types/entities.ts
export interface BaseMolecule {
  /** Molecule snowflake ID */
  id: string;
  /** Molecule name (free text) */
  name: string;
  /** Molecule short alias */
  alias: string;
  /** Mol smiles formula (optional) */
  smiles: string;
  /** Category, should be a GO Term */
  category: string[];
  /** Molecule version (free text) */
  version: string;
  /** Free comment text. */
  comments: string;
  /** Citation */
  citation: string;
  /** Information about a protein validation, model quality */
  validation: string;
  /** String version of the used command line (other parameters) */
  command_line: string;
  /** Way to create the martinized molecule (id that refers in create_way field of settings.json) */
  create_way: string;
  /** Force field version */
  force_field: string;
  /** Stringified ISO date of creation date */
  created_at: string;
  /** Molecule parent version. If string, ref to <Molecule.id> */
  parent: null | string;
  /** Tree snowflake ID. Shared between parent and children */
  tree_id: string;
  /** Hash of generated zip file attached to this module */
  hash: string;
  /** Reference to <User.id> owner/curator of this mol */
  owner: string;
  /** ID of related file containing `.itp` and `.gro`/`.pdb` files */
  files: string;
  /** Author (if fetched). */
  author?: string;
  builder_mode ?: string; 
  alternative_alias?: string[]; 
  from_stashed? : boolean; 
}

export interface ModelMolecule {
  name: string; 
  alias: string; 
  smiles: string; 
  category: string[]; 
  tree_id: string; 
  force_field?: string; 
}

export interface Molecule extends BaseMolecule {
  /** <User.id> that have approved this molecule */
  approved_by: string;
  /** Last time as ISO date the user/admin edited this molecule */
  last_update: string;
}
// ...
interface JobDoc extends CouchDoc {
  id: string; 
  jobId: string; 
  userId : string;
  date : string; 
  radius : { [atomName: string]: number }; 
  settings: JobSettings 
  type : "martinize" | "insane" | "polyply"; 
  name: string;
  update_date? : string; 
  manual_bonds_edition?: boolean; 
  comment: string; 
}

export interface ReadedJobDoc extends JobDoc { 
  files : ReadedJobFiles; 

}

export interface RawJobDoc extends JobDoc {
  files: RawJobFiles

}

export interface ReadedJobFiles {
  all_atom : FileFromHttp; 
  coarse_grained : FileFromHttp; 
  itp_files : FileFromHttp[][]; 
  top_file : FileFromHttp; 
  warnings: FileFromHttp; 
  gro?: FileFromHttp; 
}

export interface RawJobFiles { 
  all_atom : FileFromHttp; 
  coarse_grained : FileFromHttp; 
  itp_files : FileFromHttp[][]; 
  top_file : FileFromHttp;
}


export interface JobSettings {
  builder_mode : MartinizeMode; 
  ff : AvailableForceFields; 
  advanced : boolean; 
  commandline : string; 
  cter : "neutral"; 
  nter : "charged";
  sc_fix : boolean; 
  position : "backbone" | "all" | "none"
  cystein_bridge : "none" | "auto"
  elastic? : boolean; 
  use_go? : boolean; 
  ea? : string; 
  ef? : string; 
  el? : string; 
  em? : string; 
  ep? : string; 
  eu? : string; 
}

export interface FileFromHttp {
  name: string; 
  type: string; 
  content: string; 
}

interface CouchDoc {
  _id : string; 
  _rev : string; 
}